made setting resolver to return the realtime params

+++ services/truapi/truapi/types.go
@@ -104,12 +104,12 @@ type Slash struct {
 
 // Settings contains application specific settings
 type Settings struct {
-	MinClaimLength    int64    `json:"minClaimLength"`
-	MaxClaimLength    int64    `json:"maxClaimLength"`
-	MinArgumentLength int64    `json:"minArgumentLength"`
-	MaxArgumentLength int64    `json:"maxArgumentLength"`
-	MinSummaryLength  int64    `json:"minSummaryLength"`
-	MaxSummaryLength  int64    `json:"maxSummaryLength"`
+	MinClaimLength    int      `json:"minClaimLength"`
+	MaxClaimLength    int      `json:"maxClaimLength"`
+	MinArgumentLength int      `json:"minArgumentLength"`
+	MaxArgumentLength int      `json:"maxArgumentLength"`
+	MinSummaryLength  int      `json:"minSummaryLength"`
+	MaxSummaryLength  int      `json:"maxSummaryLength"`
 	MinCommentLength  int64    `json:"minCommentLength"`
 	MaxCommentLength  int64    `json:"maxCommentLength"`
 	BlockIntervalTime int64    `json:"blockIntervalTime"`
